rename: 파일명 isValidImageFile.ts를 validateImageFile.ts로 수정
