+++ src/features/user/utils/validateImageFile.ts
